+++ migration-library/src/main/kotlin/com/quadient/migration/api/dto/migrationmodel/Ref.kt
@@ -171,29 +171,3 @@ data class StringValue(val value: String) : TextContent {
     fun toModel() = StringModel(value)
     fun toDb() = StringEntity(value)
 }
-
-data class Hyperlink(
-    val url: String,
-    val displayText: String? = null,
-    val alternateText: String? = null
-) : TextContent {
-    companion object {
-        fun fromModel(model: HyperlinkModel) = Hyperlink(
-            url = model.url,
-            displayText = model.displayText,
-            alternateText = model.alternateText
-        )
-    }
-
-    fun toModel() = HyperlinkModel(
-        url = url,
-        displayText = displayText,
-        alternateText = alternateText
-    )
-
-    fun toDb() = HyperlinkEntity(
-        url = url,
-        displayText = displayText,
-        alternateText = alternateText
-    )
-}