+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ### Added
 
 - PDF tagging support for tables
+- Alternate text option for images including support in export/import mapping scripts
+- New Hyperlink text content 
 
 ### Changed